move test to utils
